added places search suggestions

+++ srv.php
@@ -13,13 +13,24 @@ switch ($_GET['cmd']) {
         }
 
         $url = 'https://api.meteo.lt/v1/places';
+        $file = 'cache/places.json';
+        
+        $content = '';
+        if (!file_exists($file) || filemtime($file) + 600 < time()) {
+            $handle = fopen($file, 'w');
+            $content = file_get_contents($url);
 
-        $content = file_get_contents($url);
+            if ($content == false) {
+                echo json_encode(['error' => true, 'error_name' => 'failed to fetch url contents']);
+                exit;
+            }
 
-        if ($content == false) {
-            echo json_encode(['error' => true, 'error_name' => 'failed to fetch url contents']);
-            break;
+            fwrite($handle, $content);
+            fclose($handle);
+        } else {
+            $content = file_get_contents($file);
         }
+
 
         if (!$places_array = json_decode($content)) {
             echo json_encode(['error' => true, 'error_name' => 'failed to fetch url contents']);
@@ -27,6 +38,9 @@ switch ($_GET['cmd']) {
         }
 
         $word = strtolower($_GET['search_phrase']);
+        if (trim($word) == '') {
+            $word = 'abc';
+        }
         $len = strlen($word);
         $matches = array();
         foreach ($places_array as $obj) {
